pyproject.toml
[tool.poetry]
name = "myteams"
version = "0.1.0"
description = ""
authors = ["qu3vipon <sdh5813@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
django-allauth = "^0.41.0"
pillow = "^7.0.0"
boto3 = "^1.12.5"
django-storages = "^1.9.1"
psycopg2-binary = "^2.8.4"
django-crispy-forms = "^1.8.1"
bs4 = "^0.0.1"
request = "^2019.4.13"
selenium = "^3.141.0"
requests = "^2.23.0"
django-extensions = "^2.2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
